requirements.txt
numpy~=2.2.6
pillow~=11.2.1